requirements.txt
Click==7.0
fastjsonschema==2.14.1
pkg-resources==0.0.0
pur==5.2.2
pytoolz==0.1.7
unpythonic==0.14.1
